% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Net.R
\name{plot_Net}
\alias{plot_Net}
\title{Plot an igraph network with customizable node sizes and edge widths}
\usage{
plot_Net(
  graph,
  label = FALSE,
  color = "#006d77",
  color_ramp = c("blue", "white", "red"),
  NA_color = "gray",
  frame.color = NULL,
  node.size.factor = 1,
  node.degree.map = TRUE,
  edge.color = "#999999",
  edge.width.factor = 1,
  edge.bw.map = TRUE,
  label.color = "#e29578",
  label.size = 1,
  layout = NULL,
  ...
)
}
\arguments{
\item{graph}{An \code{igraph} object representing the network to plot or a
data frame containing a symbolic edge list in the first two columns. Additional
columns are considered as edge attributes.}

\item{label}{Logical, or character vector. If \code{FALSE}, no node labels
are shown. If \code{TRUE}, node labels are set to vertex names or existing
labels. If a character vector of length equal to the number of vertices,
used as custom node labels.}

\item{color}{Character. Color for node fill, or existing vertex attribute name
to map node fill color. Default is \code{'#006d77'}.}

\item{color_ramp}{A character vector of colors (e.g., \code{c("blue", "white", "red")}) used to create a continuous color ramp
for mapping numeric node attributes to colors. Ignored if a fixed color is provided or if node color is categorical.
Passed to \code{\link[grDevices]{colorRampPalette}} to interpolate a gradient of colors.}

\item{NA_color}{Node color for NAs in node color mapping.}

\item{frame.color}{Character. Color for node frame, or existing vertex attribute name
to map node frame color. Default is \code{NULL}.}

\item{node.size.factor}{Numeric or character. If numeric, acts as a constant
scaling factor for node size. If character, it should be the name of a
numeric vertex attribute used to size nodes.}

\item{node.degree.map}{Logical. If \code{TRUE} and \code{node.size.factor} is numeric,
node size is mapped to vertex degree multiplied by \code{node.size.factor}.
Ignored if \code{node.size.factor} is character.}

\item{edge.color}{Character. Color of edges. Default is \code{'#999999'}.}

\item{edge.width.factor}{Numeric. Factor to scale edge widths. Default is 1.}

\item{edge.bw.map}{Logical. If \code{TRUE}, edge widths are mapped to the
log-transformed edge betweenness centrality. Otherwise, edges have uniform width.}

\item{label.color}{Character. Color of node labels. Default is \code{'#e29578'}.}

\item{label.size}{Numeric. Character expansion factor for label size. Default is 1.}

\item{layout}{Optional numeric matrix specifying vertex coordinates for layout.}

\item{...}{Additional parameters passed to \code{plot.igraph}.}
}
\value{
Invisibly returns \code{NULL}. The function produces a plot.
}
\description{
This function plots an \code{igraph} network graph with options to
customize node size based on vertex degree or a numeric vertex attribute,
set node and edge colors, label nodes, and adjust edge widths
(optionally mapped to edge betweenness centrality).
}
\examples{
g <- igraph::sample_gnp(50, 0.06, directed = FALSE)
plot_Net(g, label = TRUE, node.size.factor = 2)

}
